R6 Class representing a LimeSurvey question

R6 Class representing a LimeSurvey question

R6 Class representing a LimeSurvey question

Details

A question has at least a code and a primary language.

Public fields

code

The code of the question.

id

The identifier of the question (a unique number in a survey)

type

The question type.

lsType

The question type in LimeSurvey's format.

questionTexts

The question text(s) in all languages.

helpTexts

The question help text(s) in all languages.

relevance

The relevance.

validation

The question's validation.

language

The primary language of the question.

answerOptions

The answer options in the question.

subQuestions

The subquestions in the question.

mandatory

Whether the question is mandatory (Y or N).

other

Whether the question has an 'other' option (Y or N).

other_replace_text

If the question has an 'other' option, its label if the default label should be overwritten (multilingual).

default

The default value.

same_default

Not entirely sure what this does.

array_filter

The question code of the array filter question to apply.

cssclass

The CSS class(es) to apply to this question.

hide_tip

Whether to hide the tip (Y or N).

otherOptions

Any additional options, stored as a named list by assigning as.list(...).

Methods

Public methods

Question$new()

Question$add_answer_option()

Question$clone()

Method new()

Create a new question object.

Usage

Question$new(
  code,
  type = NULL,
  lsType = NULL,
  id = NULL,
  questionTexts = "",
  helpTexts = "",
  relevance = 1,
  validation = "",
  mandatory = "N",
  other = "N",
  other_replace_text = "",
  default = "",
  same_default = "0",
  array_filter = "",
  cssclass = "",
  hide_tip = "",
  language = "en",
  ...
)

Arguments

code

The question code.

type

The human-readable question type.

lsType

The type as LimeSurvey type (see https://manual.limesurvey.org/Question_object_types).

id

The identifier of the question (in a survey).

questionTexts

The question text(s).

helpTexts

The help text(s).

relevance

The question's relevance equation.

validation

The question's validation.

mandatory

Whether the question is mandatory (Y or N);.

other

Whether the question has an 'other' option (Y or N).

other_replace_text

If the question has an 'other' option, its label if the default label should be overwritten (multilingual).

default

The default value.

same_default

Not entirely sure what this does.

array_filter

The question code of the array filter question to apply.

cssclass

The CSS class(es) to apply to this question.

hide_tip

Whether to hide the tip (Y or N).

language

The question's primary language.

...

Any additional options, stored as a named list in the otherOptions property by assigning as.list(...).

Returns

A new Survey object.

Method add_answer_option()

Add an answer option to a question.

Usage

Question$add_answer_option(code, optionTexts)

Arguments

code

The answer option code.

optionTexts

The answer option text(s).

Returns

Invisible, the question object.

Method clone()

The objects of this class are cloneable with this method.

Usage

Question$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.