A function to conveniently process LimeSurvey labels

This function is meant to quickly parse the variable labels set by LimeSurvey. It works particularly well with dual anchor array questions, where the left and right anchors as well as the subquestions are extracted automatically.

ls_process_labels(
  dat,
  varnameRegExPairs = NULL,
  lengthToWrap = 50,
  lengthToWrapAnchors = 20,
  labelExtractionRegExPair = limonaid::opts$get(labelExtractionRegExPair),
  leftAnchorRegExPairs = limonaid::opts$get(leftAnchorRegExPairs),
  rightAnchorRegExPairs = limonaid::opts$get(rightAnchorRegExPairs)
)

Arguments

dat
The dataframe as produced by ls_import_data().

varnameRegExPairs
Pairs of regular expressions to replace in the variable names. This is useful when some pattern can be applied to the variable names to, for example, add underscores te denote different parts of the variable name. This has to be a list of character vectors that each have length 2.

lengthToWrap
At how many characters to wrap the subquestions.

lengthToWrapAnchors
At how many characters to wrap the anchors.

labelExtractionRegExPair
The regular expression pair used to extract the labels.

leftAnchorRegExPairs
The regular expression pairs to use to extract the left anchors.

rightAnchorRegExPairs
The regular expression pairs to use to extract the right anchors.

Value

A dataframe.

Details

This function processes LimeSurvey variable labels and applies regular expressions to automatically extract subquestions and left and right anchors.

Examples

### No examples provided yet; this would require data to be included,
### and that's not available yet.