R6 Class representing a LimeSurvey survey

R6 Class representing a LimeSurvey survey

R6 Class representing a LimeSurvey survey

Details

A survey has at least a title and a primary language.

Public fields

title

The title of the survey in the primary language.

primaryLanguage

The primary language of the survey.

groups

The groups in the survey.

Active bindings

get_group_ids

A list of all group ids.

Methods

Public methods

Survey$new()

Survey$add_group()

Survey$add_question()

Survey$clone()

Method new()

Create a new survey object.

Usage

Survey$new(title, primaryLanguage = "en")

Arguments

title

Title.

primaryLanguage

The primary language.

Returns

A new Survey object.

Method add_group()

Add a group to a survey object.

Usage

Survey$add_group(titles, descriptions = "", relevance = 1, random_group = "")

Arguments

titles

The group's title, either as a named character vector where each element is the group title in a different language, and every element's name is the language code; or as a single character value, in which case the survey's primary language is used.

descriptions

The group description, either as a named character vector where each element is the group description in a different language, and every element's name is the language code; or as a single character value, in which case the survey's primary language is used.

relevance

The group's relevance equation.

random_group

The group's randomization group.

Returns

Invisibly, the Survey object.

Method add_question()

Add a question to a survey object.

Usage

Survey$add_question(groupId, code, type = NULL, lsType = NULL, ...)

Arguments

groupId

The id of the group to add the question to.

code

The question code.

type

The question type.

lsType

The question type, as LimeSurvey question type.

...

Additional arguments are used to create the Question using Question$new.

Returns

Invisibly, the Survey object.

Method clone()

The objects of this class are cloneable with this method.

Usage

Survey$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.